CRUD Users Password Encryption
Users created through the manager myworkers page, or the admin allusers page, now encrypts the inputted password. (Previously the password was not encrypted on input. The log in with new users would not work as the login password field uses encryption, while there was no encryption when registering the user)